requirements.txt
azure-storage-blob
azure-functions
azure-core
azure-ai-formrecognizer